+++ Source/Core/Utility/AnyValueArray.h
@@ -1,33 +1,15 @@
 /****************************************************************************/
 /**
- *  @file AnyValueArray.h
- */
-/*----------------------------------------------------------------------------
- *
- *  Copyright (c) Visualization Laboratory, Kyoto University.
- *  All rights reserved.
- *  See http://www.viz.media.kyoto-u.ac.jp/kvs/copyright/ for details.
- *
- *  $Id: AnyValueArray.h 1707 2014-01-27 07:37:04Z naohisa.sakamoto@gmail.com $
+ *  @file   AnyValueArray.h
+ *  @author Naohisa Sakamoto
  */
 /****************************************************************************/
-#ifndef KVS__ANY_VALUE_ARRAY_H_INCLUDE
-#define KVS__ANY_VALUE_ARRAY_H_INCLUDE
-
+#pragma once
 #include <typeinfo>
 #include <utility>
-#if KVS_ENABLE_DEPRECATED
-#include <string>
-#include <sstream>
-#endif
 #include <kvs/Type>
 #include <kvs/SharedPointer>
 #include <kvs/ValueArray>
-#if KVS_ENABLE_DEPRECATED
-#include <kvs/Endian>
-#include <kvs/String>
-#include <kvs/AnyValue>
-#endif
 #include "StaticAssert.h"
 
 
@@ -85,9 +67,6 @@ public:
         case kvs::Type::TypeUInt64: return GetValue<ReturnT, kvs::UInt64>( ptr );
         case kvs::Type::TypeReal32: return GetValue<ReturnT, kvs::Real32>( ptr );
         case kvs::Type::TypeReal64: return GetValue<ReturnT, kvs::Real64>( ptr );
-#if KVS_DEPRECATED
-        case kvs::Type::TypeString: return GetValueFromString<ReturnT>( ptr );
-#endif
         default: break;
         }
         return ReturnT();
@@ -100,47 +79,9 @@ private:
         return static_cast<ReturnT>( *static_cast<const DataT*>( ptr ) );
     }
 
-#if KVS_ENABLE_DEPRECATED
-    template <typename DataT>
-    static std::string GetString( const void* ptr )
-    {
-        return kvs::String::From( *static_cast<const DataT*>( ptr ) );
-    }
-
-    template <typename ReturnT>
-    static ReturnT GetValueFromString( const void* ptr )
-    {
-        const std::string& source = *static_cast<const std::string*>( ptr );
-        return kvs::String::To<ReturnT>( source );
-    }
-#endif
 private:
     ValueGetter();
 }; // ValueGetter
-
-#if KVS_ENABLE_DEPRECATED
-template <>
-inline std::string ValueGetter::Get<std::string>( const void* ptr, kvs::Type::TypeID id )
-{
-    KVS_ASSERT( id != kvs::Type::UnknownType );
-    switch ( id )
-    {
-    case kvs::Type::TypeInt8:   return GetString<kvs::Int8>  ( ptr );
-    case kvs::Type::TypeInt16:  return GetString<kvs::Int16> ( ptr );
-    case kvs::Type::TypeInt32:  return GetString<kvs::Int32> ( ptr );
-    case kvs::Type::TypeInt64:  return GetString<kvs::Int64> ( ptr );
-    case kvs::Type::TypeUInt8:  return GetString<kvs::UInt8> ( ptr );
-    case kvs::Type::TypeUInt16: return GetString<kvs::UInt16>( ptr );
-    case kvs::Type::TypeUInt32: return GetString<kvs::UInt32>( ptr );
-    case kvs::Type::TypeUInt64: return GetString<kvs::UInt64>( ptr );
-    case kvs::Type::TypeReal32: return GetString<kvs::Real32>( ptr );
-    case kvs::Type::TypeReal64: return GetString<kvs::Real64>( ptr );
-    case kvs::Type::TypeString: return *static_cast<const std::string*>( ptr );
-    default: break;
-    }
-    return std::string();
-}
-#endif
 
 class AnyValueArrayElement
 {
@@ -306,7 +247,6 @@ private:
     size_t    m_size_of_value; ///< byte size of a value
     size_t    m_size;       ///< number of values
     kvs::Type::TypeID m_type_id;
-
     TypeInfo m_type_info;
 
 public:
@@ -387,137 +327,18 @@ public:
     }
 
     // for compatibility.
-
+    // {
     const TypeInfo* typeInfo() const
     {
         return &m_type_info;
     }
 
-
-public:
-#if KVS_ENABLE_DEPRECATED
-
-    template<typename T>
-    AnyValueArray( const T* values, const size_t size )
-    {
-        *this = AnyValueArray( kvs::ValueArray<T>( values, size ) );
-    }
-
-    template<typename T>
-    AnyValueArray( const std::vector<T>& values )
-    {
-        *this = AnyValueArray( kvs::ValueArray<T>( values ) );
-    }
-
-    template<typename T>
-    void* allocate( const size_t size )
-    {
-        this->deallocate();
-        *this = AnyValueArray( kvs::ValueArray<T>( size ) );
-        return this->data();
-    }
-
-    bool isEmpty() const
-    {
-        return this->empty();
-    }
-
-    void swapByte()
-    {
-        switch ( this->typeID() )
-        {
-        case kvs::Type::TypeInt8:   kvs::Endian::Swap( static_cast<kvs::Int8*  >( this->data() ), this->size() ); break;
-        case kvs::Type::TypeUInt8:  kvs::Endian::Swap( static_cast<kvs::UInt8* >( this->data() ), this->size() ); break;
-        case kvs::Type::TypeInt16:  kvs::Endian::Swap( static_cast<kvs::Int16* >( this->data() ), this->size() ); break;
-        case kvs::Type::TypeUInt16: kvs::Endian::Swap( static_cast<kvs::UInt16*>( this->data() ), this->size() ); break;
-        case kvs::Type::TypeInt32:  kvs::Endian::Swap( static_cast<kvs::Int32* >( this->data() ), this->size() ); break;
-        case kvs::Type::TypeUInt32: kvs::Endian::Swap( static_cast<kvs::UInt32*>( this->data() ), this->size() ); break;
-        case kvs::Type::TypeInt64:  kvs::Endian::Swap( static_cast<kvs::Int64* >( this->data() ), this->size() ); break;
-        case kvs::Type::TypeUInt64: kvs::Endian::Swap( static_cast<kvs::UInt64*>( this->data() ), this->size() ); break;
-        case kvs::Type::TypeReal32: kvs::Endian::Swap( static_cast<kvs::Real32*>( this->data() ), this->size() ); break;
-        case kvs::Type::TypeReal64: kvs::Endian::Swap( static_cast<kvs::Real64*>( this->data() ), this->size() ); break;
-        default: break;
-        }
-    }
-
-    template<typename T>
-    void shallowCopy( const kvs::ValueArray<T>& values )
-    {
-        *this = AnyValueArray( values );
-    }
-
-    void shallowCopy( const AnyValueArray& other )
-    {
-        *this = other;
-    }
-
-    template<typename T>
-    void deepCopy( const T* values, const size_t size )
-    {
-        *this = AnyValueArray( kvs::ValueArray<T>( values, size ) );
-    }
-
-    const void* pointer() const
-    {
-        return this->data();
-    }
-
-    void* pointer()
-    {
-        return this->data();
-    }
-
-    template<typename T>
-    const T* pointer() const
-    {
-        KVS_STATIC_ASSERT( is_supported<T>::value, "not supported" );
-        KVS_ASSERT( this->check_type<T>() );
-        return static_cast<const T*>( this->data() );
-    }
-
-    template<typename T>
-    T* pointer()
-    {
-        KVS_STATIC_ASSERT( is_supported<T>::value, "not supported" );
-        KVS_ASSERT( this->check_type<T>() );
-        return static_cast<T*>( this->data() );
-    }
-
-    //template<typename T>
-    //T& at( const size_t index )
-    //{
-    //    KVS_STATIC_ASSERT( is_supported<T>::value, "not supported" );
-    //    KVS_ASSERT( index < this->size() );
-    //    KVS_ASSERT( this->check_type<T>() );
-    //    return static_cast<T*>( this->data() )[ index ];
-    //}
-
-    //template<typename T>
-    //const T& at( const size_t index ) const
-    //{
-    //    KVS_STATIC_ASSERT( is_supported<T>::value, "not supported" );
-    //    KVS_ASSERT( index < this->size() );
-    //    KVS_ASSERT( this->check_type<T>() );
-    //    return static_cast<const T*>( this->data() )[ index ];
-    //}
-
-    template<typename T>
-    T to( const size_t index ) const
-    {
-        KVS_STATIC_ASSERT( is_supported<T>::value, "not supported" );
-        return ( *this )[ index ].to<T>();
-    }
-
-    void deallocate();
-
-#else
     template<typename T>
     void allocate( const size_t size )
     {
         *this = AnyValueArray( kvs::ValueArray<T>( size ) );
     }
-
-#endif
+    // }
 
     void release();
 
@@ -544,11 +365,8 @@ template <> struct AnyValueArray::is_supported<kvs::UInt32> : kvs::temporal::tru
 template <> struct AnyValueArray::is_supported<kvs::UInt64> : kvs::temporal::true_type {};
 template <> struct AnyValueArray::is_supported<kvs::Real32> : kvs::temporal::true_type {};
 template <> struct AnyValueArray::is_supported<kvs::Real64> : kvs::temporal::true_type {};
-#if KVS_ENABLE_DEPRECATED
-template <> struct AnyValueArray::is_supported<std::string> : kvs::temporal::true_type {};
-#endif
-} // end of namespace kvs
 
+} // end of namespace kvs
 
 namespace std
 {
@@ -560,5 +378,3 @@ inline void swap( kvs::AnyValueArray& lhs, kvs::AnyValueArray& rhs )
 }
 
 } // std
-
-#endif // KVS__ANY_VALUE_ARRAY_H_INCLUDE